mcom: Program ALWAYS_MISC0 register to make SoC reboot possible
Once watchdog has been triggered the SoC is rebooted and bootrom
_warm_reset branch is used for SoC booting. It's assumed that
ALWAYS_MISC0 contains valid jump address to continue booting. So if
we want to boot the SoC in the same way as it's done after power on
we should set ALWAYS_MISC0 to start address of bootrom _cold_reset
branch.

Change-Id: Id247e2af3983b1e16111aa53a3b1cff22ff58d1a