designware_spi: Add support for GPIO chip select
Designware SPI driver is modified to support GPIO chip select.
This is necessary to communicate with SPI FLASH devices. If
hardware chip select is used, transfer is not working properly
because of line toggling in the duration of one transfer cycle.
Used GPIO number is passed via "cs-gpios" property of the device
tree node, describing the SPI master device.
Driver Model for GPIO drivers should be enabled for building
the driver.

Change-Id: I064aa3f692a46151eb8565e1c7fe2c0c86cb0d78