mcom: Reimplement ddr_init()
According to specification for DDRMC reset deassertion must
be synchronous to DDRMC clock. Actually DDRMC reset deassertion for in
MCom-02 is synchronous to XTI_24, which differs from DDRMC clock
if CPLL is enabled.

To avoid this problem (rf#1963) reset for DDRMC and PHY is deasserted
now at 24 МНz (CPLL is off). CPLL is turned on later. Frequency changing
for ACDLL is done when ACDLL is in reset mode.

This prevents CPU hanging during DDR read operation at boot.

Issue: rf#1963
Change-Id: If900e471bf71ba44491441f4f99fec236f011dd4