mcom: Store environment variables in SPI flash
The environment storage is moved from MMC to SPI flash.

The SPI flash partitioning has been changed to allocate the space for
the environment storage. The first erase sector is still used for the
SPL image. The second is reserved for the environment storage. The
U-Boot image begins from the third sector.

The configuration define CONFIG_SYS_MONITOR_LEN is removed, because
it is not used for the current environment placement.

Change-Id: I5133a5d58e54145bd0c7f7ec02a93a1f1101f5ab