mcom: Disable the early malloc() pool
The early malloc() pool is reserved below a top of the stack to
provide simple malloc() support for board_init_f(). The simple
malloc() is required by Driver Model to operate before U-Boot
relocation.

The early malloc() pool should not be allocated for SPL, because it
is started in the limited memory area, but U-Boot configuration does
not allow to set up the pool for SPL and U-Boot separately.

Driver Model support is not required before U-Boot relocation, and
the early malloc() pool can be disabled to free memory for SPL.

Change-Id: Ia1ba18c6a2aa514c88a98577d482b1df23edbdf4