mcom: Add u-boot.mcom to default target
This is required to build u-boot.mcom when 'make all' is invoked.
Currently there is not such option in kconfig.

Alternative way to build u-boot.mcom as a default target is to add
it to ALL variable when CONFIG_ARCH_MCOM is defined in top-level
Makefile. I don't know what is preferred by U-Boot maintainers, but
I think it is better not to flood top-level Makefile.

Change-Id: Ie48905bf76bf64780c74c50241868a3f93b1f39a
Signed-off-by: Anton Leontiev <aleontiev@elvees.com>