ARM: dts: mcom: Use recommended values for L2 cache latencies
Now that recommended values are added to the user manual, it is
better to use them instead of old values that were obtained
experimentally.

Note that in DeviceTree three values specify read, write and setup
latencies in that order. In registers they are write, read and setup
latencies from high to low nibble.

Issue: rf#2031
Change-Id: I23163c2192cdf2e7db933925473f4ad5f0f32c9e
Signed-off-by: Anton Leontiev <aleontiev@elvees.com>