Add mcom02-nand driver
The driver adds support for MCom-02 NFC. It's based on Linux version
of arasan_nfc driver v2.99. The source has been copied from Linux
mcom02/v4.4.111.5 with a few changes to make it work in U-Boot.

MDMA mode is not supported. The only supported mode is PIO.

Issue: #MCOM02SW-854
Change-Id: I50efe18983fec9a4b48b3aace0dcf5c99fe887d4