mcom: usb: gadget: Add DWC2 driver probing
The commit adds the support of USB Device mode and is an initial
step of implementing support of USB Mass Storage and Device Firmware
Update features.

Since the DWC2 OTG Device driver has no uclass, it should be probed
manually. And thus one gets a properly initialized low level driver
for the operation of the G_DNL driver which underlies the functioning
of UMS and DFU.

Issue: #MCOM02SW-756
Change-Id: Ib23bcf324e49acb4a9cee833b73d663e0b13c8c4