mcom: Add "video" parameter to the Linux kernel command line
Because LCD panels have no detection mechanism, they can be selected
as default even if we don't want use it.

Linux command line parameter "video" can be used to select proper
connector as default and to create a correct framebuffer for it.
This parameter is defined only for Salute-EL24PM boards and will be
stored in configuration header.

Issue: #MCOM02SW-821
Change-Id: I38a45f65ba89add8b48069f5d13dfc89910d082b