mcom: Rename CONFIG_TIMER_CLK_FREQ macro
The CONFIG prefix forces to interpret the option as a configurable
option which should be provided in Kconfig menu or in the
config-whitelist. This requirement is not fully correct for the
timer clock frequency because this value does not imply a manual
adjustment and it is calculated automatically for each single board.

Removing the prefix allows one to build U-Boot without specifying a
build target.

Issue: #MCOM02SW-1048
Change-Id: Ic36a7640bbbe14b558dc75afa1dac43120eb05de