ARM: dts: mcom: Decrease SPI0.1 frequency to 20 MHz
Sometimes bit flips occur when data is read from memory
to controller at frequency 40 MHz and oscilloscope probe
is connected to MISO signal or MCom-02 chip is heated.

When reading from memory at frequency 20 MHz bit flips
errors do not occur.

It is important that the actual frequency was reduced
from 36 MHz to 18 MHz due to the fact that it is
impossible to get 40 MHz and 20 MHz from the value of
the frequency of the SPI clock bus 144 MHz with an
integer divider.

Issue: #MCOM02SW-1252
Change-Id: I6ad3d18d59ec5be0b885d8d2b21ee8920bf30e16