ARM: dts: mcom: Reserve memory in DDR0 and DDR1 for VPU
VPU decoder must use memory in DDR0 and DDR1 in order to implement
workaround to prevent DDR-VDMA deadlock (issue rf#1382).

It is required to reserve contiguous page aligned memory regions in
DDR0 and DDR1 with size equal to

  S = roundup(Sr, PAGE_SIZE) + roundup(Se, PAGE_SIZE),

where

* Sr = roundup(W, 16) * roundup(H, 16) * 3 / 2 + roundup(W, 16) * 12
  is the size of buffer for reference frame;
* Se = W * H * 2 is the size of buffer for encoded frame.

The patch reserves 8 MiB in DDR0 and 8 MiB in DDR1 for VPU which are
enough for 1920x1072 resolution. The memory is reserved only for
Salute EL24PM* boards which have two DDRs.

The regions are excluded from the common usage by the kernel. The
iomem information (/proc/iomem) will show that the System RAM is less
than the amount of memory on the board.

Issue: rf#1382
Issue: #MCOM02SW-1573
Change-Id: I06010d947c1523c346846545140e53c5ad577153