mcom: Add ECAM02DM boot environment
ECAM02DM boot system from NAND using UBIFS and OverlayFS.

The standard legacy_ubifs0 assume that kernel is on boot
volume that is not true for ECAM02DM. For this reason
ecam02dm_boot command was added.

Also, standard boot commands assume that rootfs should be
mounted as read-write. So, the kernel argument "rw" was
placed in additional rootfs_options variable.

Change-Id: If777002d6a1f62667088457e0ef66fe881edbd92