mcom03: Fix 'fdtfile' environment variable
Since 'mcom03.h' is required at the configuration stage before 'dt.h'
is created, we cannot use 'DEVICE_TREE' to define 'fdtfile' variable.

Due to this limitation, Linux will boot with the default DeviceTree.
To change this, it is needed to directly set DeviceTree in
'extlinux.conf'.

Issue: #MCOM03BUP-225
Change-Id: I31748d357479989380152ad9c39df313b46bb57c