mcom03: ecam03: Don't set 'fdtfile' variable
'fdtfile' variable is set to 'mcom03-ecam03bl-r1.0.dtb' for all
ECAM03* boards. This is inconvenient because it is required to
manually redefine 'fdtfile' for boards other than ECAM03BL r1.0.
This commit removes the default setting of 'fdtfile'. Instead,
variables 'soc' and 'board' will be used to select DTB file for
Linux.

Issue: #MCOM03SW-2281
Change-Id: I1f935e9beb93bb04e56fed4a52514dc15fc09a67