mcom03: Import factory settings from eMMC boot0
This patch allows to import factory settings from the text file
"uboot-factory.env". It is required that the file has key-value pairs
separated with the '\n' symbol. It is expected that the file is located
in an eMMC boot0 partition. The partition theoretically can be formated
to any file system compatible to eMMC device and supported by U-Boot but
tested only with ext4.

If factory settings have 'factory_wp' variable then whole eMMC boot0
partition will be write protected with Power-On mode. This behavior can
be disabled by enabling the CONFIG_MCOM03_EMMC_BOOT_WP_DISABLED option.
The option is useful in case of build for development purpose or it is
required to modify factory settings.

If an eMMC device, any supported file system or the "uboot-factory.env"
file on eMMC boot0 device are not found then a warning message will be
displayed and boot process will continue without factory settings.

Issue: #MCOM03SW-2381
Depends-On: https://gerrit.elvees.com/c/mcom03/sbl/+/38053
Change-Id: I0ed1cfbc30b5fe392bda90b5c2299f1f791f5dcd