mcom03: dts: Rename DTS for ECAM03* r2.0 boards
Currently at first boot U-Boot reads 'compatible' from U-Boot DTB and
tries to find the corresponding Linux DTB. For ECAM03* r2.0, the problem
is that U-Boot uses a common DTB for all ECAM03* r2.0, which has no
analogue in Linux. As a result at first boot it is required to manually
override the 'board' variable in U-Boot terminal, but this is inconvenient
as it requires UART connection.

Common for all ECAM03* r2.0 boards Linux DTB is added in [1]. This commit
renames U-Boot DTS to match Linux DTB file name.

[1] https://review.elvees.com/c/mcom03/linux/+/1714

Issue: #MCOM03SW-2382
Change-Id: I7b360248bc389721e5620ee51876ae99e510768e