i2c: designware_i2c: Add ability to control STOP bit
The current driver always forcibly sends a STOP bit at the end of each
data transfer, which is not compliant with the I2C standard. Some
devices terminate the transaction and reset the register address for
read/write operations.

The I2C DW driver has been fixed for any values of
IC_EMPTYFIFO_HOLD_MASTER_EN hardware parameter. The change allows
clients to control the sending of STOP bit themselves. This is done by
reducing the delay when filling the TX buffer.

With very slow CPU operations this may not be sufficient and the
problem will be reproduced.

Issue: #MCOM03SW-2473
Change-Id: I919edda355df60057f2f719896a0eda975bfa4b0