checkpatch: Add support to ignore specific check
The change is imported from [1].

The patch allows to ignore the specific check based on the git trailers
of commit message providing a list of checkpatch tests to ignore
delimited by comma.

Example of usage:
  Checkpatch-Ignore: <tests to ignore 0>, <tests to ignore 1>, ...

For example it can be used to ignore the checkpatch issue such as:

  WARNING:PREFER_IF: Use 'if (IS_ENABLED(CONFIG...))' instead of '#if or #ifdef' where possible

using the following git trailers:

  Checkpatch-Ignore: PREFER_IF

By default checkpath doesn't provide a message type. To show a message
type in output messages the '--show-types' command line argument has to
be provided in checkpatch call.

[1] https://review.openocd.org/c/openocd/+/6169

Issue: #FORTESW-58
Depends-On: https://gerrit.elvees.com/c/infra/zuul-jobs/+/50469
Change-Id: Id5d2a96eda3fe3cc60674d16535f2f3783d8d121
Signed-off-by: Mikhail Konovalov <mkonovalov@elvees.com>