mcom03: factory-settings: Check return code
We have to check return code of `detect_board_name` to properly catch
cases when board name is not set.

Also we have to check return code of `load_factory_settings` to properly
catch errors of loading of the factory settings.

Also CONFIG_MCOM03_DISABLE_FACTORY enabled has to be considered as
valid case and should not raise the error.

Issue: #MCOM03SW-1510
Change-Id: I94161916f45fccb9d8ee99e7b116b958a32514e0
Signed-off-by: Mikhail Konovalov <mkonovalov@elvees.com>