mcom03: Use constant mask in FIELD_PREP
The "pad_set_bits" function is used only in "pad_set_ctl" where macro
FIELD_PREP is used with the constant mask. It is optimized in case of
release build and no error is occurred but in case of debug build or
one with instrumentation, the optimization doesn't take effect and
the following issue is occurred:

  error: call to '__compiletime_assert_0' declared with attribute error: FIELD_PREP: mask is not constant

To fix the issue let's combine functions into "hsp_emac_pad_set_ctl"
one with the constant mask related to HSP subsystem.

Also COMMIT_LOG_LONG_LINE is ignored here as error messages could be
long and this usually ignored in case of automatic tests in our
projects.

Issue: #MCOM03SW-3347
Depends-On: https://gerrit.elvees.com/c/mcom03/tests/+/52836
Checkpatch-Ignore: COMMIT_LOG_LONG_LINE
Change-Id: I383601b75b0c885ac30d592e7c4c09579b0cae1a
Signed-off-by: Mikhail Konovalov <mkonovalov@elvees.com>