services: mcom03: Add bootstage SiP services
The bootstage is a library designed to analyze boot time. It uses
a monotonic system timer and allow to store timestamps with stage
identifier to memory in the form of records.

The commit adds SiP calls to record timestamp of U-Boot start event
(board_early_init_f) and timestamp of OS run event (board_prep_linux).

Issue: #MCOM03SW-3347
Change-Id: I9b24faed07c19487b7821165d94b301a71e55d55
Signed-off-by: Mikhail Konovalov <mkonovalov@elvees.com>