mcom02: do not use __stringify(FDTFILE)
For CONFIG_EXTRA_ENV_SETTINGS the FDTFILE is taken from the
CONFIG_DEFAULT_DEVICE_TREE .._defconfig file. In .._defconfig the
value of CONFIG_DEFAULT_DEVICE_TREE value is a string enclosed in double
quotes. When using stringify these double quotes will be carried over to
the resulting string, which will cause the following problem:

  Retrieving file: /boot/elvees/"mcom02-salute-el24pm2-r1.0-1.1-om1-r1.2".dtb
  ** File not found /boot/elvees/"mcom02-salute-el24pm2-r1.0-1.1-om1-r1.2".dtb **

We just need to pass the FDTFILE value as is.

At the same time, we need to convert DEVICE_TREE to string if it is
passed from command `make DEVICE_TREE=<device_tree_file_name>` and do
this in the arch/arm/Makefile by adding escaped double quotes.

Issue: #MCOM02SW-2403
Change-Id: I3980948c961dc85a301bb1ed0f7559ffe17d884b