factory-settings: Disable console if necessary
If environment variable 'disable_console' is set to "1" (directly or by
means of 'factory_disable_console' variable), the console (in & out)
will be disabled after the factory settings are applied.

It is useful in the following cases to prevent to enter in the console
during boot of Linux kernel:

* secure boot - to restrict access to U-Boot commands to decrease
  attack surface;
* automatic test - to prevent occasionally stopping of Linux kernel
  boot process for the tests involving reboot.

The 'factory_disable_console' variable has more priority than
'disable_console' one.

Issue: #MCOM03SW-3539
Change-Id: I0c23ee1458f587bc1591a732351bd8c4812c5c66
Signed-off-by: Mikhail Konovalov <mkonovalov@elvees.com>